abstract CubQMC{Lattice,Sobol}G to CubQMCLDG class
